refactor: move ir tests into ir/tests/ subfolder
